Add presence status for auto create channels, Rename incorrect namespace OilPrice.
