+++ DisbotNext/ExternalServices/OilPriceChecker/OilPriceWebScraping.cs
@@ -7,7 +7,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace DisbotNext.ExternalServices.OildPriceChecker
+namespace DisbotNext.ExternalServices.OilPriceChecker
 {
     public class OilPriceWebScraping : IOilPriceChecker
     {
