+++ DisbotNext/Program.cs
@@ -1,7 +1,7 @@
 ﻿using DisbotNext.Common.Configurations;
 using DisbotNext.DiscordClient;
 using DisbotNext.ExternalServices.CovidTracker;
-using DisbotNext.ExternalServices.OildPriceChecker;
+using DisbotNext.ExternalServices.OilPriceChecker;
 using DisbotNext.Infrastructure.Common;
 using DisbotNext.Infrastructures.Sqlite;
 using Hangfire;
@@ -75,7 +75,7 @@ namespace DisbotNext
                     services.AddHttpClient();
                     services.AddDbContext<DisbotDbContext, SqliteDbContext>();
                     services.AddHangfire(config =>
-                    {
+                   
                         config.UseColouredConsoleLogProvider();
                         config.UseLiteDbStorage();
                     });
